fix a compilaiton error on Windows
git-svn-id: https://root.cern.ch/svn/root/branches/dev@33973 27541ba8-7e3a-0410-8455-c3a389f83636

+++ roostats/roofit/roofitcore/src/RooAbsPdf.cxx
@@ -611,7 +611,7 @@ Double_t RooAbsPdf::getLogVal(const RooArgSet* nset) const
 
     logEvalError("getLogVal() top-level p.d.f evaluates to zero") ;
 
-    return log(0);
+    return log((double)0);
   }
   return log(prob);
 }
